refactor: reorganize extension layout
