+++ extensions/_shared/auth.ts
