+++ extensions/provider/quota-store.ts
@@ -1,8 +1,8 @@
 import type { AuthStorage } from "@earendil-works/pi-coding-agent";
-import { getNeuralwattApiKey } from "../../lib/env";
-import { fetchQuotas } from "../../lib/neuralwatt-api";
-import type { NeuralwattQuotas } from "../../types/quota-api";
-import { parseQuotaHeaders } from "../../types/quota-events";
+import { parseQuotaHeaders } from "../../src/events";
+import { fetchQuotas } from "../../src/lib/neuralwatt-api";
+import type { NeuralwattQuotas } from "../../src/types/quota-api";
+import { getNeuralwattApiKey } from "../_shared/auth";
 
 export function buildQuotasFromHeaders(
   headers: Record<string, string>,
