+++ src/config/defaults.ts
@@ -1,0 +1,9 @@
+import type { ResolvedNeuralwattConfig } from "./types";
+
+export const DEFAULT_CONFIG: ResolvedNeuralwattConfig = {
+  quotaCommand: true,
+  quotaWarnings: true,
+  subBarIntegration: true,
+  includeLegacyModelIds: false,
+  includeHiddenModels: false,
+};
