+++ src/config/index.ts
@@ -1,0 +1,4 @@
+export { DEFAULT_CONFIG } from "./defaults";
+export { configLoader } from "./loader";
+export { migrations } from "./migration";
+export type { NeuralwattConfig, ResolvedNeuralwattConfig } from "./types";
